Final todo ok

+++ src/controllers/postsController.js
@@ -1,24 +1,47 @@
-const { selectAll, selectById, insertPost, updatePostById, deleteById } = require("../models/postsModel");
+const { selectAll, selectById, getPostsByAuthor, insertPost, updatePostById, deleteById } = require("../models/postsModel");
+const { post } = require("../routes/apiRoutes");
 
 const getAllPosts = async (req, res, next) => {
+    console.log("entra0");
     try {
         const [result] = await selectAll()
         res.json(result);
     } catch (error) {
         next(error);
     }
-}
+};
 
 
 const getById = async (req, res, next) => {
-    const { postId } = req.params;
+    let id = req.params.id;
+    console.log("parametros:" ,id);
     try {
-        const post = await selectById(postId);
-        res.json(post);
+        const post = await selectById(id);
+        if (!post) {
+            return res.status(404).json({ message: 'Post no encontrado' });
+        }
+        console.log(post);
+        res.json(post);  
     } catch (error) {
-        next(error);
+        next(error);  
     }
-}
+};
+
+const selectPostByAuthor = async (req, res, next) => {
+    let id = req.params.id;
+    console.log("parametros:" ,id);
+    try {
+        const post = await getPostsByAuthor(id);
+        if (!post) {
+            return res.status(404).json({ message: 'Post no encontrado' });
+        }
+        console.log(post);
+        res.json(post);  
+    } catch (error) {
+        next(error);  
+    }
+};
+
 
 //las funciones dentro de controlador son las de modelos
 const createPost = async (req, res, next) => {
@@ -34,7 +57,7 @@ const createPost = async (req, res, next) => {
     } catch (error) {
         next(error);
     }
-}
+};
 
 const updatePost = async (req, res, next) => {
     const { postId } = req.params;
@@ -46,7 +69,7 @@ const updatePost = async (req, res, next) => {
     } catch (error) {
         next(error);
     }
-}
+};
 
 const deletePost= async (req, res, next) => {
     const { postId } = req.params;
@@ -58,8 +81,8 @@ const deletePost= async (req, res, next) => {
     } catch (error) {
         next(error);
     }
-}
+};
 
 module.exports = {
-    getAllPosts, getById, createPost, updatePost, deletePost 
+    getAllPosts, getById, selectPostByAuthor, createPost, updatePost, deletePost
 }
